+++ src/egon/data/datasets/scenario_parameters/__init__.py
@@ -190,7 +190,7 @@ class ScenarioParameters(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="ScenarioParameters",
-            version="0.0.7",
+            version="0.0.8",
             dependencies=dependencies,
             tasks=(
                 create_table,